Missed moving the file upload status item to the filebrowser package.
