+++ packages/filebrowser-extension/src/index.ts
@@ -22,14 +22,15 @@ import { IDocumentManager } from '@jupyterlab/docmanager';
 import {
   FileBrowserModel,
   FileBrowser,
+  FileUpload,
   IFileBrowserFactory
 } from '@jupyterlab/filebrowser';
-
-import { fileUploadStatus } from './uploadstatus';
 
 import { Launcher } from '@jupyterlab/launcher';
 
 import { Contents } from '@jupyterlab/services';
+
+import { IStatusBar } from '@jupyterlab/statusbar';
 
 import { IIterator, map, reduce, toArray } from '@phosphor/algorithm';
 
@@ -123,6 +124,36 @@ const shareFile: JupyterLabPlugin<void> = {
   id: '@jupyterlab/filebrowser-extension:share-file',
   requires: [IFileBrowserFactory],
   autoStart: true
+};
+
+/**
+ * A plugin providing file upload status.
+ */
+export const fileUploadStatus: JupyterLabPlugin<void> = {
+  id: '@jupyterlab/filebrowser-extension:file-upload-status',
+  autoStart: true,
+  requires: [IStatusBar, IFileBrowserFactory],
+  activate: (
+    app: JupyterLab,
+    statusBar: IStatusBar,
+    browser: IFileBrowserFactory
+  ) => {
+    const item = new FileUpload({
+      tracker: browser.tracker
+    });
+
+    statusBar.registerStatusItem(
+      '@jupyterlab/filebrowser-extension:file-upload-status',
+      {
+        item,
+        align: 'middle',
+        isActive: () => {
+          return !!item.model && item.model.items.length > 0;
+        },
+        activeStateChanged: item.model.stateChanged
+      }
+    );
+  }
 };
 
 /**
